Split simulation_t from simif.h
