+++ sim/midas/src/main/cc/bridges/simulation.cc
@@ -1,0 +1,87 @@
+// See LICENSE for license details.
+
+#include "simulation.h"
+#include "bridges/timing.h"
+#include "simif.h"
+
+#include <cassert>
+#include <cinttypes>
+#include <cstdio>
+
+void simulation_t::record_start_times() {
+  start_hcycle = sim.actual_hcycle();
+  start_time = timestamp();
+}
+
+void simulation_t::record_end_times() {
+  end_time = timestamp();
+  end_tcycle = sim.actual_tcycle();
+  end_hcycle = sim.actual_hcycle();
+}
+
+void simulation_t::print_simulation_performance_summary() {
+  // Must call record_start_times and record_end_times before invoking this
+  // function
+  assert(start_hcycle != -1 && end_hcycle != 0 && "simulation not executed");
+
+  const uint64_t hcycles = end_hcycle - start_hcycle;
+  const double sim_time = diff_secs(end_time, start_time);
+  const double sim_speed = ((double)end_tcycle) / (sim_time * 1000.0);
+  const double measured_host_frequency =
+      ((double)hcycles) / (sim_time * 1000.0);
+  const double fmr = ((double)hcycles / end_tcycle);
+
+  fprintf(stderr, "\nEmulation Performance Summary\n");
+  fprintf(stderr, "------------------------------\n");
+  fprintf(stderr, "Wallclock Time Elapsed: %.1f s\n", sim_time);
+  // Provide enough sig-figs to let the report be useful in RTL sim
+  fprintf(stderr, "Host Frequency: ");
+  if (measured_host_frequency > 1000.0) {
+    fprintf(stderr, "%.3f MHz\n", measured_host_frequency / 1000.0);
+  } else {
+    fprintf(stderr, "%.3f KHz\n", measured_host_frequency);
+  }
+
+  fprintf(stderr, "Target Cycles Emulated: %" PRIu64 "\n", end_tcycle);
+  fprintf(stderr, "Effective Target Frequency: ");
+  if (sim_speed > 1000.0) {
+    fprintf(stderr, "%.3f MHz\n", sim_speed / 1000.0);
+  } else {
+    fprintf(stderr, "%.3f KHz\n", sim_speed);
+  }
+  fprintf(stderr, "FMR: %.2f\n", fmr);
+  fprintf(stderr,
+          "Note: The latter three figures are based on the fastest "
+          "target clock.\n");
+}
+
+int simulation_t::execute_simulation_flow() {
+  simulation_init();
+
+  record_start_times();
+  fprintf(stderr, "Commencing simulation.\n");
+  const int exit_code = simulation_run();
+  fprintf(stderr, "\nSimulation complete.\n");
+  record_end_times();
+
+  simulation_finish();
+
+  const bool timeout = simulation_timed_out();
+
+  if (exit_code != 0) {
+    fprintf(stderr,
+            "*** FAILED *** (code = %d) after %" PRIu64 " cycles\n",
+            exit_code,
+            end_tcycle);
+  } else if (timeout) {
+    fprintf(stderr,
+            "*** FAILED *** simulation timed out after %" PRIu64 " cycles\n",
+            end_tcycle);
+  } else {
+    fprintf(stderr, "*** PASSED *** after %" PRIu64 " cycles\n", end_tcycle);
+  }
+
+  print_simulation_performance_summary();
+
+  return timeout ? EXIT_FAILURE : exit_code;
+}
