+++ sim/src/main/cc/fasedtests/fasedtests_top.h
@@ -7,6 +7,7 @@
 #include "bridges/bridge_driver.h"
 #include "bridges/fpga_model.h"
 #include "bridges/peek_poke.h"
+#include "bridges/simulation.h"
 #include "firesim/systematic_scheduler.h"
 
 class fasedtests_top_t : public systematic_scheduler_t, public simulation_t {
